requirements.txt
django>=1.5,<1.6
-e git://github.com/vhf/python-markdown-linkfix@master#egg=Markdown-2.3.1
bleach==1.2.2
South==0.8.2
pytz==2013d
Pillow
unidecode
django-crispy-forms>=1.4.0
crispy-forms-foundation>=0.2.3
django-simple-captcha>=0.4.0
django-email-obfuscator>=0.1.1
django-floppyforms>=1.1
django-mptt==0.6.0
simplejson==3.3.0
django-libs==1.16
-e git://github.com/pinax/django-mailer.git@919805e25fc18b06487b25414d4e5d5e05a92108#egg=django_mailer-dev
django-storages
-e git://github.com/jhgaylor/tldrio-python-client.git@master#egg=tldrio
-e git://github.com/vhf/django-taggit.git@master#egg=django-taggit
-e git://github.com/vhf/django-mptt-comments.git@master#egg=django-mptt-comments
python-memcached
psycopg2
django-celery
